+++ lib/Frontend/CMakeLists.txt
@@ -19,8 +19,8 @@ add_dependencies(swiftFrontend
 target_link_libraries(swiftFrontend PRIVATE
   swiftAST
   swiftConstExtract
+  swiftIDE
   swiftSIL
-  swiftMigrator
   swiftOption
   swiftSILGen
   swiftSILOptimizer
