+++ lib/IDETool/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 add_swift_host_library(swiftIDETool STATIC
   CompileInstance.cpp
+  CompilerInvocation.cpp
   CompletionInstance.cpp
   DependencyChecking.cpp
   )
@@ -9,6 +10,8 @@ target_link_libraries(swiftIDETool PRIVATE
   swiftAST
   swiftDriver
   swiftFrontend
-  swiftFrontendTool)
+  swiftFrontendTool
+  clangAST
+  clangFrontend)
 
 set_swift_llvm_is_available(swiftIDETool)
